testing new isdark

+++ inc/sys_config.h
@@ -326,7 +326,7 @@ void displayDim(void);
 void displayNormal(void);
 void setIOpin(struct ALARM_SYSTEM_S *sys, uint8_t level);
 uint8_t getIOpin(struct ALARM_SYSTEM_S *sys);
-uint8_t isDark(uint8_t mode);
+uint16_t isDark(uint8_t mode);
 void saveByte(uint32_t offset, uint8_t ebyte);
 uint32_t bytesToint(uint8_t *bytes);
 void intTobytes(uint8_t *bytes, uint32_t intVal);
